Add SOURCE_EXCLUDE setting to filter the source account's tweets
SOURCE_EXCLUDE is a regular expression - any tweets from the source
account that match it will not be passed on to the Markov generation and
so will not be used to produce the bot's tweets. My intention behind
this feature is to prevent your bot from "learning" inappropriate words
or phrases, thereby producing inappropriate tweets.