Use random.randint() to make the guess, rather than choosing from a range()
Also: when ODDS = 0, we completely skip the randomisation step. This
fixes a bug in the previous implementation, which was that, since
range(0) is an empty list, random.choice(range(0)) threw an exception
rather than always returning zero.

ODDS = 0 and ODDS = 1 now have identical behaviour. This is correct
behaviour, since the bot should tweet with probability 1/ODDS and 1/1
probability means "always".